requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
beautifulsoup4==4.13.3
bertopic==0.17.0
blis==1.3.0
catalogue==2.0.10
cbor==1.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
emoji==2.14.1
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
executing==2.2.0
filelock==3.13.1
fsspec==2024.6.1
hdbscan==0.8.40
huggingface-hub==0.30.2
idna==3.10
ijson==3.3.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
ir_datasets @ git+https://github.com/allenai/ir_datasets.git@3d657a7b9f159382776763ee093fc6ef2d7c6581
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
keybert==0.9.0
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
llvmlite==0.44.0
lxml==5.3.2
lz4==4.4.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.34.1
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.61.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.7
plotly==6.0.1
preshed==3.0.9
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pyautocorpus==0.1.12
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.2
trec-car-tools==2.6
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.7
unlzw3==0.2.3
urllib3==2.4.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
widgetsnbextension==4.0.14
wrapt==1.17.2
zlib-state==0.1.9
